refactor(tests): split process runner suite
