+++ Tests/ComposeCoreTests/ComposeOrchestratorTests.swift
@@ -3339,93 +3339,6 @@ struct ComposeOrchestratorTests {
         }
     }
 
-    @Test("process runner captures stdout stderr status input env and cwd")
-    func processRunnerCapturesProcessDetails() async throws {
-        let directory = FileManager.default.temporaryDirectory
-        let script = "printf \"%s:%s\" \"$PROCESS_RUNNER_VALUE\" \"$(pwd)\"; cat; printf err >&2"
-        let result = try await ProcessRunner().run(
-            "/bin/sh",
-            ["-c", script],
-            workingDirectory: directory,
-            environment: ["PROCESS_RUNNER_VALUE": "ok"],
-            input: Data(" input".utf8)
-        )
-
-        #expect(result.succeeded)
-        #expect(
-            result.stdout == "ok:\(directory.path) input"
-                || result.stdout == "ok:/private\(directory.path) input"
-        )
-        #expect(result.stderr == "err")
-    }
-
-    @Test("recording runner captures command environment")
-    func recordingRunnerCapturesCommandEnvironment() async throws {
-        let runner = RecordingRunner()
-
-        _ = try await runner.run("/usr/bin/env", ["true"], environment: ["SAMPLE": "value"])
-
-        let command = try #require(runner.commands.first)
-        #expect(command.environment == ["SAMPLE": "value"])
-    }
-
-    @Test("recording runner captures command input")
-    func recordingRunnerCapturesCommandInput() async throws {
-        let runner = RecordingRunner()
-        let input = Data("payload".utf8)
-
-        _ = try await runner.run("/usr/bin/env", ["true"], input: input)
-
-        let command = try #require(runner.commands.first)
-        #expect(command.input == input)
-    }
-
-    @Test("process runner reports status when inheriting terminal IO")
-    func processRunnerReportsStatusWhenInheritingTerminalIO() async throws {
-        let result = try await ProcessRunner().run(
-            "/bin/sh",
-            ["-c", "exit 7"],
-            workingDirectory: nil,
-            environment: nil,
-            io: .inherited
-        )
-
-        #expect(result.status == 7)
-        #expect(result.stdout.isEmpty)
-        #expect(result.stderr.isEmpty)
-    }
-
-    @Test("process runner drains large stdout and stderr while process runs")
-    func processRunnerDrainsLargeOutputWhileProcessRuns() async throws {
-        let result = try await ProcessRunner().run(
-            "/bin/sh",
-            [
-                "-c",
-                """
-                python3 - <<'PY'
-                import sys
-                sys.stdout.write("o" * 262144)
-                sys.stdout.flush()
-                sys.stderr.write("e" * 262144)
-                sys.stderr.flush()
-                PY
-                """,
-            ]
-        )
-
-        #expect(result.succeeded)
-        #expect(result.stdout.count == 262_144)
-        #expect(result.stderr.count == 262_144)
-    }
-
-    @Test("process runner reports nonzero status")
-    func processRunnerReportsNonzeroStatus() async throws {
-        let result = try await ProcessRunner().run("/bin/sh", ["-c", "printf nope >&2; exit 9"])
-
-        #expect(!result.succeeded)
-        #expect(result.status == 9)
-        #expect(result.stderr == "nope")
-    }
 }
 
 private extension CommandResult {
